Crate first_application

This crate contains a simple routine to start the first application (or set of applications).

This should be invoked at or towards the end of the kernel initialization procedure.

Important Dependency Note

In general, Theseus kernel crates cannot depend on application crates. However, this crate is a special exception in that it directly loads and runs the first application crate.

Thus, it’s safest to ensure that first application crate is always included in the build by specifying it as a direct dependency here.

Currently, that crate is applications/shell, but if it changes, we should change that dependendency in this crates Cargo.toml manifest.

Functions

start

Starts the first applications that run in Theseus by creating a new “default” application namespace and spawning the first application Task(s).